Normalize line endings for .bat files
When updating the Gradle wrapper's .bat file, the line endings were
replaced with ^M instead of keeping the existing CRLF ending.

A few samples were already using a .gitattributes file to explicitely
set the line endings for .bat files to CRLF. Move the .gitattributes
file to the root folder and delete the sample specific files.